Progress on bakebot with new images and text.
This includes what should be (almost!) the rest of the bakebot images, and some modified text.

I changed the behavior so there is no special max water constraint; it's just max ingredient, but whether you see the huge piles of flour or the flood depend on what you maxxed out to get that alert.

Added the winner image, but the win condition does not seem to be triggered, so that's my next task.