Fixes #21. (Wonky win conditions.)
This fixes issue #21, and adds what I think is a nice win condition
state. In short, IF all conditions are set to appropriate values, AND
you bake something, it will present you with the winner image and badge.
If you bake anything in that state, it will just give you the winner
image, making it so that you can't "miss" the fact that you solved the
puzzle. However, if you move one of the constraints out (e.g., set a
minimum to <0) then it will hide the win image. This lets people
experiment with the bot, but makes it clear when they're right.

I think we are ready for alpha testers.